Needed to use rmv_non_stalling to get it to work.
I added USE_RMV and changed the code so it would work with using
rmv or rmv_non_stalling exclusively. It appears my stall routine
is incorrect and just waiting for pTail->pNext to be non NULL is
not sufficient. Or I'm doing it incorrectly.

Added a bunch of debug especially using of gTick to be able to
generate a sequence number and help understand the code flow.