Account for msgs and cmds processed.
Now msgs_processed is the total number of messages removed from fifo's and
cmds_processed is the number of command messages processed by the
clients.

So cmds_processed represents a measure of "real" work by the system and
msgs_processed is a measure of time to add and remove a message.

Here are some numbers, not neccessarily accurate but should be useful
for detecting algorithm changes:
```
wink@wink-desktop:~/prgs/test-mpscfifo-intrusive (master)
$ ./test 12 10000000 65536
test client_count=12 loops=10000000 msg_count=65536
     1 7fef8658b700  multi_thread_msg:+client_count=12 loops=10000000 msg_count=65536
     2 7fef8658b700  multi_thread_msg: msgs_processed=1792982654 msgs_sent=77888428 no_msgs_count=42111572
     3 7fef8658b700  startup=0.043570
     4 7fef8658b700  looping=30.239260
     5 7fef8658b700  disconneting=0.138639
     6 7fef8658b700  stopping=0.005271
     7 7fef8658b700  complete=0.009876
     8 7fef8658b700  processing=30.393s
     9 7fef8658b700  cmds_per_sec=29482574
    10 7fef8658b700  ns_per_cmd=33.9ns
    11 7fef8658b700  msgs_per_sec=58993186
    12 7fef8658b700  ns_per_msg=17.0ns
    13 7fef8658b700  total=30.437
    14 7fef8658b700  multi_thread_msg:-error=0

Success
```